OJORISE-154 FEAT: 요금제 클릭 시, 차트로 스크롤 되게 변경.
